加入了connector和define
加入了connector（树突）的c文件
加入了define（定义）的文件

+++ connector.h
@@ -4,7 +4,7 @@
 
 typedef struct connector
 {
-    int weight;
+    double weight;
     int local_id;               //本地ID由columnid和cellid组成。
     int target_id;              //目标ID由columnid和cellid组成。
 }Connector, *pConnector;
